Use layout and page transitions instead of changing page transition
In the Nuxt 1.x version of the codebase, the transition hook was used to
change the page transition to a fade transition if users were moving
directly between two non-index routes.

Nuxt 3 now has layout transitions in addition to page transitions. Using
layout transitions since this models the desired behavior a bit better.